requirements.txt
customtkinter
pillow
pyperclip
numpy
scikit-learn